Add modules js
