+++ src/index.js
@@ -1,224 +1,7 @@
 import './style.css';
-
-class CardListClass {
-  constructor (container, arrayData) {
-    this.container = container
-    this.arrayData = arrayData
-    this.render()
-  }
-
-  addCard (nameCard, linkCard) {
-    this.container.appendChild(Card.create(nameCard, linkCard))
-  }
-}
-
-class Card {
-  constructor (name, linkPic, cardId, authorId) {
-    this.name = name
-    this.linkPic = linkPic
-    this.cardId = cardId
-    this.authorId = authorId
-    this.element = Card.create(this.name, this.linkPic, this.cardId, this.authorId)
-  }
-
-  static create (name, linkPic, cardId, authorId) {
-    const myId = 'e02e6975b9ecfc5e75a38b25'
-
-    const card = document.createElement('div')
-    // Создаем основной div
-    const newCard = document.createElement('div')
-    newCard.classList.add('place-card')
-
-    card.appendChild(newCard)
-
-    // Создаем div для фона и кнопки "удалить карту"
-    const imageContainer = document.createElement('div')
-    imageContainer.classList.add('place-card__image')
-    newCard.appendChild(imageContainer)
-    imageContainer.setAttribute('style', `background-image: url(${linkPic})`)
-
-    if (authorId === myId) {
-      const deleteBtn = document.createElement('button')
-      deleteBtn.classList.add('place-card__delete-icon')
-      deleteBtn.setAttribute('card-id', `${cardId}`)
-      newCard.setAttribute('author-id', `${authorId}`)
-      imageContainer.appendChild(deleteBtn)
-    }
-
-    // Создаем div для контейнера имени и кнопки "лайк"
-    const descriptionContainer = document.createElement('div')
-    descriptionContainer.classList.add('place-card__description')
-    newCard.appendChild(descriptionContainer)
-
-    // Создаем имя карточки
-    const newName = document.createElement('h3')
-    newName.classList.add('place-card__name')
-    newName.textContent = `${name}`
-    descriptionContainer.appendChild(newName)
-
-    // Создаем кнопку "лайк"
-    const likeBtn = document.createElement('button')
-    likeBtn.classList.add('place-card__like-icon')
-    descriptionContainer.appendChild(likeBtn)
-
-    return newCard
-  }
-
-  static like (element) {
-    element.classList.toggle('place-card__like-icon_liked')
-  }
-
-  static remove () {
-    const card = event.target.closest('.place-card')
-    card.parentElement.removeChild(card)
-  }
-}
-
-class Popup {
-  constructor (element, openClassName, closeElement) {
-    this.element = element
-    this.openClassName = openClassName
-    closeElement.addEventListener('click', () => this.close())
-  }
-
-  open () {
-    this.element.classList.add(this.openClassName)
-  }
-
-  close () {
-    this.element.classList.remove(this.openClassName)
-  }
-}
-
-class Api {
-  constructor (authorization) {
-    this.authorization = authorization
-  }
-
-  getInitialCards (baseUrl, method) {
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`
-      }
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  getUserInfo (baseUrl, method) {
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`
-      }
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  changeUserInfo (baseUrl, method) {
-    editBtn.textContent = 'Загрузка...'
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        name: `${inputUserName.value}`,
-        about: `${inputUserAbout.value}`
-      })
-    })
-
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  postCard (baseUrl, method, nameCard, linkCard) {
-    postBtn.textContent = 'Загрузка...'
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        name: `${nameCard}`,
-        link: `${linkCard}`
-      })
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  likeSum (baseUrl, method) {
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`
-      }
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  changeAvatar (baseUrl, method) {
-    postAvatar.textContent = 'Загрузка...'
-    return fetch(`http://95.216.175.5/cohort1/${baseUrl}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`,
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        avatar: `${inputAvatarUrl.value}`
-      })
-    })
-
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-
-  deleteCard (cardId, method) {
-    return fetch(`http://95.216.175.5/cohort1/cards/${cardId}`, {
-      method: `${method}`,
-      headers: {
-        authorization: `${this.authorization}`
-      }
-    })
-      .then((res) => {
-        if (res.ok) {
-          return res.json()
-        }
-        return Promise.reject(`Ошибка: ${res.status}`)
-      })
-  }
-}
+import Card from './blocks/place-card/place-card';
+import Api from './scripts/api';
+import Popup from './blocks/popup/popup';
 
 // Переменные
 // Карточки
